Add login and sign up

+++ pages/deleteUser.php
@@ -1,12 +1,20 @@
 <?php
-require_once("./class/class.User.php");
+session_start();
 
-if($_GET["user_id"]){
-    $objUser = new User();
-    $objUser->user_id = $_GET["user_id"];
-    $objUser->DeleteUser();
+if (!isset($_SESSION["user_id"])) {
+    header("location: index.php?p=signin");
+    exit();
 
-    echo "<script>alert('$objUser->message');</script>";
-    echo "<script>window.location = 'index.php?p=listUser';</script>";
+} else {
+    require_once ("./class/class.User.php");
+
+    if ($_GET["user_id"]) {
+        $objUser = new User();
+        $objUser->user_id = $_GET["user_id"];
+        $objUser->DeleteUser();
+
+        echo "<script>alert('$objUser->message');</script>";
+        echo "<script>window.location = 'index.php?p=listUser';</script>";
+    }
 }
- ?>
+?>
